refactor(rust): 使用 derive 宏简化 Default 实现
- 使用 #[derive(Default)] 替代手动 Default 实现
- 为 enum 添加 #[default] 属性指定默认值
- 减少样板代码，提高可读性

+++ src-tauri/src/commands/context.rs
@@ -307,7 +307,7 @@ impl ContextMemoryStore {
         let selected: Vec<ContextEntry> = entries
             .into_iter()
             .take_while(|e| {
-                total_tokens += e.estimated_tokens as u32;
+                total_tokens += e.estimated_tokens;
                 total_tokens <= max_tokens
             })
             .collect();
